requirements.txt
# Core async
aiohttp>=3.9,<4
websockets>=12,<14

# Database
duckdb>=1.1,<2

# Cache
redis>=5,<6

# Web dashboard
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
jinja2>=3.1,<4

# Polymarket SDK (for later phases)
py-clob-client>=0.1

# Utilities
python-dotenv>=1.0,<2
structlog>=24,<26
orjson>=3.10,<4
